+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 
 ### Fixed
 
+- Creating a repo with `auto_init: false` and `gh_pages_build_type: "disabled"` is now supported.
 - Avoided unnecessary GitHub API calls when getting the `default_branch` or `id` of a repository.
 - Detect errors during an automatic `apply` operation and add a corresponding comment to the pull request.
 - Support showing dialog windows when using operation `web-login`.